add actual screens for desktop
