+++ composeApp/src/commonMain/kotlin/screens/auth/RegisterScreen.kt
@@ -17,14 +17,13 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
 import com.arkivanov.decompose.extensions.compose.subscribeAsState
+import component.CempButton
 import component.RegisterComponent
 import dev.icerock.moko.resources.compose.stringResource
-import theme.AppTheme
 import theme.Theme
-import component.CempButton
-import component.CempProgressBar
-import component.CempText
-import component.CempTextField
+import ui.component.CempProgressBar
+import ui.component.CempText
+import ui.component.CempTextField
 import utils.StringResHelper
 import com.cemp.SharedRes.strings as stringsRes
