+++ composeApp/src/commonMain/kotlin/ui/component/CempButton.kt
@@ -1,4 +1,4 @@
-package component
+package ui.component
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.PaddingValues
@@ -10,7 +10,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import theme.AppTheme
 import theme.Theme
 
 @Composable
