+++ composeApp/src/commonMain/kotlin/ui/component/CempProgressBar.kt
@@ -1,4 +1,4 @@
-package component
+package ui.component
 
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.runtime.Composable
