+++ composeApp/src/commonMain/kotlin/ui/component/CempTextField.kt
@@ -1,4 +1,4 @@
-package component
+package ui.component
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
