+++ composeApp/src/commonMain/kotlin/ui/component/ErrorBanner.kt
@@ -1,27 +1,26 @@
-package component
+package ui.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
+import com.cemp.SharedRes.strings
+import dev.icerock.moko.resources.compose.stringResource
 import theme.Theme
 
 @Composable
-fun ProgressBarBanner(modifier: Modifier = Modifier) {
+fun ErrorBanner(modifier: Modifier = Modifier) {
     Column(
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally,
         modifier = modifier.fillMaxSize().background(Theme.colors.mainBackgroundColor)
     ) {
-        CircularProgressIndicator(
-            modifier = Modifier.size(48.dp),
-            color = Theme.colors.blueTextColor,
+        Text(
+            text = stringResource(strings.feature_auth_unknown_error)
         )
     }
 }
