+++ composeApp/src/commonMain/kotlin/ui/component/TeamBanner.kt
